\name{adductize_formula}
\alias{adductize_formula}
\title{Adductize Formula}
\description{
Given a molecular formula and an adduct name, return the precursor m/z. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
adductize_formula(formula, adduct, verbose)
}
\arguments{
\item{formula}{Argument formula.}

\item{adduct}{Argument adduct.}

\item{verbose}{Argument verbose.}

}
\value{
Returns double.
}
\details{
Generated documentation for adductize_formula.
}
