\name{adductize_peptide}
\alias{adductize_peptide}
\title{Adductize Peptide}
\description{
Given a peptide sequence and an adduct name, return the precursor m/z. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
adductize_peptide(peptideSequence, adduct, verbose)
}
\arguments{
\item{peptideSequence}{Argument peptideSequence.}

\item{adduct}{Argument adduct.}

\item{verbose}{Argument verbose.}

}
\value{
Returns double.
}
\details{
Generated documentation for adductize_peptide.
}
