\name{DI_slice_library}
\alias{DI_slice_library}
\title{Di Slice Library}
\description{
Direct infusion mass spectrometry (DI) analysis pipeline function for MS/MS-all type DIMS analysis. Slice library based on ms2 segments. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
DI_slice_library(ms2_ranges, lipid_library, is_mark_unique_as_diagnostic, debug)
}
\arguments{
\item{ms2_ranges}{Argument ms2_ranges.}

\item{lipid_library}{Argument lipid_library.}

\item{is_mark_unique_as_diagnostic}{Argument is_mark_unique_as_diagnostic.}

\item{debug}{Argument debug.}

}
\value{
Returns List.
}
\details{
Generated documentation for DI_slice_library.
}
