\name{envelope_dist_peptide}
\alias{envelope_dist_peptide}
\title{Envelope Dist Peptide}
\description{
Given a peptide sequence, return the exact mass. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
envelope_dist_peptide(peptideSequence, threshold, debug)
}
\arguments{
\item{peptideSequence}{Argument peptideSequence.}

\item{threshold}{Argument threshold.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for envelope_dist_peptide.
}
