\name{expand_redundant_fragments_ms3_lib}
\alias{expand_redundant_fragments_ms3_lib}
\title{Expand Redundant Fragments Ms3 Lib}
\description{
Given an MS3 library data frame containing redundant fragments, expand' these redundant fragments into separate rows, with all other columns duplicated. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
expand_redundant_fragments_ms3_lib(ms3_lib, debug)
}
\arguments{
\item{ms3_lib}{Argument ms3_lib.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for expand_redundant_fragments_ms3_lib.
}
