\name{export_msp_lipids_library}
\alias{export_msp_lipids_library}
\title{Export Msp Lipids Library}
\description{
Given a DataFrame that originally derived from an imported msp library, write the contents of the DataFrame to an msp file. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
} debugging
}
\usage{
export_msp_lipids_library(mspLibraryPath, mspLibrary, observed_intensities, num_digits, debug)
}
\arguments{
\item{mspLibraryPath}{Argument mspLibraryPath.}

\item{mspLibrary}{Argument mspLibrary.}

\item{observed_intensities}{Argument observed_intensities.}

\item{num_digits}{Argument num_digits.}

\item{debug}{Argument debug.}

}
\value{
Returns int.
}
\details{
Generated documentation for export_msp_lipids_library.
}
