\name{get_eic}
\alias{get_eic}
\title{Get Eic}
\description{
This function exposes the C++ implementation of the get_eic routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
get_eic(samples, mzmin, mzmax, rtmin, rtmax, mslevel, smoothWindow, debug)
}
\arguments{
\item{samples}{Argument samples.}

\item{mzmin}{Argument mzmin.}

\item{mzmax}{Argument mzmax.}

\item{rtmin}{Argument rtmin.}

\item{rtmax}{Argument rtmax.}

\item{mslevel}{Argument mslevel.}

\item{smoothWindow}{Argument smoothWindow.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for get_eic.
}
