\name{get_last_rts}
\alias{get_last_rts}
\title{Get Last Rts}
\description{
This function exposes the C++ implementation of the get_last_rts routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
get_last_rts(samples, verbose, debug)
}
\arguments{
\item{samples}{Argument samples.}

\item{verbose}{Argument verbose.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for get_last_rts.
}
