\name{get_multi_file_consensus_spectrum}
\alias{get_multi_file_consensus_spectrum}
\title{Get Multi File Consensus Spectrum}
\description{
This function exposes the C++ implementation of the get_multi_file_consensus_spectrum routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
get_multi_file_consensus_spectrum(scans_table, params, verbose, debug)
}
\arguments{
\item{scans_table}{Argument scans_table.}

\item{params}{Argument params.}

\item{verbose}{Argument verbose.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for get_multi_file_consensus_spectrum.
}
