\name{get_scan_metadata}
\alias{get_scan_metadata}
\title{Get Scan Metadata}
\description{
This function exposes the C++ implementation of the get_scan_metadata routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
get_scan_metadata(sample_file, verbose, debug)
}
\arguments{
\item{sample_file}{Argument sample_file.}

\item{verbose}{Argument verbose.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for get_scan_metadata.
}
