\name{import_msp_lipids_library}
\alias{import_msp_lipids_library}
\title{Import Msp Lipids Library}
\description{
Given an msp library, return a DataFrame containing relevant summary information It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
import_msp_lipids_library(mspLibraryPath, totalNumFragments, is_include_SMILES, is_reformat_to_lipidmaps_2020, num_digits, is_prefer_file_summarizations, debug)
}
\arguments{
\item{mspLibraryPath}{Argument mspLibraryPath.}

\item{totalNumFragments}{Argument totalNumFragments.}

\item{is_include_SMILES}{Argument is_include_SMILES.}

\item{is_reformat_to_lipidmaps_2020}{Argument is_reformat_to_lipidmaps_2020.}

\item{num_digits}{Argument num_digits.}

\item{is_prefer_file_summarizations}{Argument is_prefer_file_summarizations.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for import_msp_lipids_library.
}
