\name{maldesi_create_modified_mzML}
\alias{maldesi_create_modified_mzML}
\title{Maldesi Create Modified Mzml}
\description{
corresponding sections of the mzML file adjusted appropriately, so that mzML parsing programs can interpret the data correctly. Immediately precedes a binary block - in the next iteration, the mz array of the appropriate scan will be encoded into binary. the line is passed along into the output file. Immediately precedes a binary block - in the next iteration, the intensity array of the appropriate scan will be encoded into binary. the line is passed along into the output file. Find the appropriate scan for writing binary data, then pass the data along into the output file. values are written ...
}
\usage{
maldesi_create_modified_mzML(sample_file, modified_sample_file, modification_params, verbose, debug)
}
\arguments{
\item{sample_file}{A parameter specifying the sample file.}

\item{modified_sample_file}{A parameter specifying the modified sample file.}

\item{modification_params}{A parameter specifying the modification params.}

\item{verbose}{A logical value indicating whether to print progress and informative messages.}

\item{debug}{A logical value indicating whether to enable detailed debugging output.}

}
\value{
This function does not return a value (void). It is typically called for its side effects, such as writing to a file.
}
\details{
This function is part of the Rcpp-based backend for the mzkit package. It performs operations related to maldesi create modified mzml.
The implementation is written in C++ for performance and exposed to R via Rcpp.
}
