\name{mark_fragments_ms2_lib}
\alias{mark_fragments_ms2_lib}
\title{Mark Fragments Ms2 Lib}
\description{
Given an MS2 library data frame containing redundant fragments, expand these redundant fragments into separate rows, with all other columns duplicated. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
mark_fragments_ms2_lib(ms2_lib, frags_to_mark, marker, debug)
}
\arguments{
\item{ms2_lib}{Argument ms2_lib.}

\item{frags_to_mark}{Argument frags_to_mark.}

\item{marker}{Argument marker.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for mark_fragments_ms2_lib.
}
