\name{mzk_get_isotope_parameters}
\alias{mzk_get_isotope_parameters}
\title{Mzk Get Isotope Parameters}
\description{
* @brief Given various parameter values, return a formatted string as produced by IsotopeParameters::encode() isotope enumeration isotope extraction only consider max natural abundance error if the isIgnoreNaturalAbundance flag is TRUE - from MAVEN 1 diff iso specific
}
\usage{
mzk_get_isotope_parameters(params, debug)
}
\arguments{
\item{params}{A parameter specifying the params.}

\item{debug}{A logical value indicating whether to enable detailed debugging output.}

}
\value{
A character string representing the processed output, file path, or status message.
}
\details{
This function is part of the Rcpp-based backend for the mzkit package. It performs operations related to mzk get isotope parameters.
The implementation is written in C++ for performance and exposed to R via Rcpp.
}
