\name{mzkitcpp-package}
\alias{mzkitcpp-package}
\alias{mzkitcpp}
\docType{package}
\title{mzkitcpp: C++ Backend for mzkit}
\description{
  Fast implementations of mass spectrometry manipulation and analysis functions, written in C++.
  It implements performance-critical steps such as direct infusion (DI) mass spectrometry 
  processing, spectra alignment, library searching, and isotopic envelope extraction. 
  The functionality is exposed to R via Rcpp.
  This function depends on the open-source maven_core repository
  (https://github.com/eugenemel/maven_core), where many algorithms are implemented.
}
\author{
  Maintainer: \email{phillipseitzer@calicolabs.com}
}
\keyword{package}
