\name{precursor_mass}
\alias{precursor_mass}
\title{Precursor Mass}
\description{
Given a vector of string formulas, return a vector of compound monoisotopic masses. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
precursor_mass(formula, adduct, num_digits, debug)
}
\arguments{
\item{formula}{Argument formula.}

\item{adduct}{Argument adduct.}

\item{num_digits}{Argument num_digits.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for precursor_mass.
}
