\name{predict_formula}
\alias{predict_formula}
\title{Predict Formula}
\description{
Given a vector of string formulas, return a vector of compound monoisotopic masses. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
predict_formula(mz, adduct, ppm, legal_atom_counts, debug)
}
\arguments{
\item{mz}{Argument mz.}

\item{adduct}{Argument adduct.}

\item{ppm}{Argument ppm.}

\item{legal_atom_counts}{Argument legal_atom_counts.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for predict_formula.
}
