\name{qqq_peaks}
\alias{qqq_peaks}
\title{Qqq Peaks}
\description{
This function exposes the C++ implementation of the qqq_peaks routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
qqq_peaks(mzml_file, query_transitions, params, debug, verbose)
}
\arguments{
\item{mzml_file}{Argument mzml_file.}

\item{query_transitions}{Argument query_transitions.}

\item{params}{Argument params.}

\item{debug}{Argument debug.}

\item{verbose}{Argument verbose.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for qqq_peaks.
}
