\name{smoothed_series}
\alias{smoothed_series}
\title{Smoothed Series}
\description{
Given a data vector and various other parameters, perform smoothing. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
smoothed_series(data, types, windowSizes, debug)
}
\arguments{
\item{data}{Argument data.}

\item{types}{Argument types.}

\item{windowSizes}{Argument windowSizes.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for smoothed_series.
}
