\name{adductize_exact_mass}
\alias{adductize_exact_mass}
\title{Adductize Exact Mass}
\description{
Given an exact mass and an adduct name, return the precursor m/z. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
adductize_exact_mass(exact_mass, adduct, verbose)
}
\arguments{
\item{exact_mass}{Argument exact_mass.}

\item{adduct}{Argument adduct.}

\item{verbose}{Argument verbose.}

}
\value{
Returns double.
}
\details{
Generated documentation for adductize_exact_mass.
}
