\name{DI_encoded_search_params}
\alias{DI_encoded_search_params}
\title{Di Encoded Search Params}
\description{
Direct infusion mass spectrometry (DI) analysis pipeline function for MS/MS-all type DIMS analysis. This function exposes the C++ implementation of the DI_encoded_search_params routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
DI_encoded_search_params(search_params, debug)
}
\arguments{
\item{search_params}{Argument search_params.}

\item{debug}{Argument debug.}

}
\value{
Returns String.
}
\details{
Generated documentation for DI_encoded_search_params.
}
