\name{DI_file_info}
\alias{DI_file_info}
\title{Di File Info}
\description{
Direct infusion mass spectrometry (DI) analysis pipeline function for MS/MS-all type DIMS analysis. This function exposes the C++ implementation of the DI_file_info routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
DI_file_info(samples, debug)
}
\arguments{
\item{samples}{Argument samples.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for DI_file_info.
}
