\name{DI_ms2_range_id}
\alias{DI_ms2_range_id}
\title{Di Ms2 Range Id}
\description{
Direct infusion mass spectrometry (DI) analysis pipeline function for MS/MS-all type DIMS analysis. Convert a precursor mz to a sample-specific DI MS2 range value (vector) It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
DI_ms2_range_id(di_ms2_ranges_table, prec_mzs, debug)
}
\arguments{
\item{di_ms2_ranges_table}{Argument di_ms2_ranges_table.}

\item{prec_mzs}{Argument prec_mzs.}

\item{debug}{Argument debug.}

}
\value{
Returns IntegerVector.
}
\details{
Generated documentation for DI_ms2_range_id.
}
