\name{DI_ms3_peakgroups_and_peaks}
\alias{DI_ms3_peakgroups_and_peaks}
\title{Di Ms3 Peakgroups And Peaks}
\description{
Direct infusion mass spectrometry (DI) analysis pipeline function for MS/MS-all type DIMS analysis. This function exposes the C++ implementation of the DI_ms3_peakgroups_and_peaks routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
DI_ms3_peakgroups_and_peaks(ms3_search_results_no_frags, first_peak_id, first_group_id, num_peakgroups, num_peaks, search_name, library_name, verbose, debug)
}
\arguments{
\item{ms3_search_results_no_frags}{Argument ms3_search_results_no_frags.}

\item{first_peak_id}{Argument first_peak_id.}

\item{first_group_id}{Argument first_group_id.}

\item{num_peakgroups}{Argument num_peakgroups.}

\item{num_peaks}{Argument num_peaks.}

\item{search_name}{Argument search_name.}

\item{library_name}{Argument library_name.}

\item{verbose}{Argument verbose.}

\item{debug}{Argument debug.}

}
\value{
Returns List.
}
\details{
Generated documentation for DI_ms3_peakgroups_and_peaks.
}
