\name{DI_peakgroups_and_peaks}
\alias{DI_peakgroups_and_peaks}
\title{Di Peakgroups And Peaks}
\description{
Direct infusion mass spectrometry (DI) analysis pipeline function for MS/MS-all type DIMS analysis. returns list of tables in mzrollDB- It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
DI_peakgroups_and_peaks(di_search_results_no_frags, first_peak_id, first_group_id, num_peakgroups, search_name, library_name, verbose, debug)
}
\arguments{
\item{di_search_results_no_frags}{Argument di_search_results_no_frags.}

\item{first_peak_id}{Argument first_peak_id.}

\item{first_group_id}{Argument first_group_id.}

\item{num_peakgroups}{Argument num_peakgroups.}

\item{search_name}{Argument search_name.}

\item{library_name}{Argument library_name.}

\item{verbose}{Argument verbose.}

\item{debug}{Argument debug.}

}
\value{
Returns List.
}
\details{
Generated documentation for DI_peakgroups_and_peaks.
}
