\name{DI_summarized_compounds}
\alias{DI_summarized_compounds}
\title{Di Summarized Compounds}
\description{
Direct infusion mass spectrometry (DI) analysis pipeline function for MS/MS-all type DIMS analysis. This function exposes the C++ implementation of the DI_summarized_compounds routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
DI_summarized_compounds(di_search_results, search_lib, ms2_ranges, adducts_file, debug)
}
\arguments{
\item{di_search_results}{Argument di_search_results.}

\item{search_lib}{Argument search_lib.}

\item{ms2_ranges}{Argument ms2_ranges.}

\item{adducts_file}{Argument adducts_file.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for DI_summarized_compounds.
}
