\name{exact_mass}
\alias{exact_mass}
\title{Exact Mass}
\description{
Given a molecular formula, return the exact mass. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
exact_mass(formula, verbose)
}
\arguments{
\item{formula}{Argument formula.}

\item{verbose}{Argument verbose.}

}
\value{
Returns double.
}
\details{
Generated documentation for exact_mass.
}
