\name{exact_mass_peptide}
\alias{exact_mass_peptide}
\title{Exact Mass Peptide}
\description{
Given a peptide sequence, return the exact mass. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
exact_mass_peptide(peptideSequence, verbose)
}
\arguments{
\item{peptideSequence}{Argument peptideSequence.}

\item{verbose}{Argument verbose.}

}
\value{
Returns double.
}
\details{
Generated documentation for exact_mass_peptide.
}
