\name{expand_redundant_fragments_ms2_lib}
\alias{expand_redundant_fragments_ms2_lib}
\title{Expand Redundant Fragments Ms2 Lib}
\description{
Given an MS2 library data frame containing redundant fragments, expand these redundant fragments into separate rows, with all other columns duplicated. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
expand_redundant_fragments_ms2_lib(ms2_lib, debug)
}
\arguments{
\item{ms2_lib}{Argument ms2_lib.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for expand_redundant_fragments_ms2_lib.
}
