\name{get_background_subtracted_scan_data}
\alias{get_background_subtracted_scan_data}
\title{Get Background Subtracted Scan Data}
\description{
This function exposes the C++ implementation of the get_background_subtracted_scan_data routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
get_background_subtracted_scan_data(sample_file, target_scan_num, params, verbose, debug)
}
\arguments{
\item{sample_file}{Argument sample_file.}

\item{target_scan_num}{Argument target_scan_num.}

\item{params}{Argument params.}

\item{verbose}{Argument verbose.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for get_background_subtracted_scan_data.
}
