\name{get_consensus_spectrum}
\alias{get_consensus_spectrum}
\title{Get Consensus Spectrum}
\description{
This function exposes the C++ implementation of the get_consensus_spectrum routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
get_consensus_spectrum(sample_file, scan_nums, params, verbose, debug)
}
\arguments{
\item{sample_file}{Argument sample_file.}

\item{scan_nums}{Argument scan_nums.}

\item{params}{Argument params.}

\item{verbose}{Argument verbose.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for get_consensus_spectrum.
}
