\name{get_scan_data}
\alias{get_scan_data}
\title{Get Scan Data}
\description{
This function exposes the C++ implementation of the get_scan_data routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
get_scan_data(sample_file, scan_num, verbose, debug)
}
\arguments{
\item{sample_file}{Argument sample_file.}

\item{scan_num}{Argument scan_num.}

\item{verbose}{Argument verbose.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for get_scan_data.
}
