\name{lipidmaps_2020_compound_names}
\alias{lipidmaps_2020_compound_names}
\title{Lipidmaps 2020 Compound Names}
\description{
Given an msp library, return a DataFrame containing relevant summary information It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
lipidmaps_2020_compound_names(ms2_lib, debug)
}
\arguments{
\item{ms2_lib}{Argument ms2_lib.}

\item{debug}{Argument debug.}

}
\value{
Returns void.
}
\details{
Generated documentation for lipidmaps_2020_compound_names.
}
