\name{maldesi_isotopic_envelope_finder}
\alias{maldesi_isotopic_envelope_finder}
\title{Maldesi Isotopic Envelope Finder}
\description{
clean up
}
\usage{
maldesi_isotopic_envelope_finder(sample_file, params, verbose, debug)
}
\arguments{
\item{sample_file}{A parameter specifying the sample file.}

\item{params}{A parameter specifying the params.}

\item{verbose}{A logical value indicating whether to print progress and informative messages.}

\item{debug}{A logical value indicating whether to enable detailed debugging output.}

}
\value{
A data frame containing the results of the mass spectrometry analysis or data extraction.
}
\details{
This function is part of the Rcpp-based backend for the mzkit package. It performs operations related to maldesi isotopic envelope finder.
The implementation is written in C++ for performance and exposed to R via Rcpp.
}
