\name{maldesi_isotopic_envelope_finder}
\alias{maldesi_isotopic_envelope_finder}
\title{Maldesi Isotopic Envelope Finder}
\description{
This function exposes the C++ implementation of the maldesi_isotopic_envelope_finder routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
maldesi_isotopic_envelope_finder(sample_file, params, verbose, debug)
}
\arguments{
\item{sample_file}{Argument sample_file.}

\item{params}{Argument params.}

\item{verbose}{Argument verbose.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for maldesi_isotopic_envelope_finder.
}
