\name{maldesi_search}
\alias{maldesi_search}
\title{Maldesi Search}
\description{
This function exposes the C++ implementation of the maldesi_search routine to R. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
maldesi_search(sample_file, library, search_params, verbose, debug)
}
\arguments{
\item{sample_file}{Argument sample_file.}

\item{library}{Argument library.}

\item{search_params}{Argument search_params.}

\item{verbose}{Argument verbose.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for maldesi_search.
}
