\name{match_id_candidates_rcpp}
\alias{match_id_candidates_rcpp}
\title{Match Id Candidates Rcpp}
\description{
This function is a reimplementation of the matching part of clamdb<matching.R>::match_id_candidates The inputs and outputs are the same as in the R code. --> $groupId (numeric vector) --> $parentGroupId
} Rcout << "-------------...
}
\usage{
match_id_candidates_rcpp(peakgroup_mzrange, standards_data_ions)
}
\arguments{
\item{peakgroup_mzrange}{Argument peakgroup_mzrange.}

\item{standards_data_ions}{Argument standards_data_ions.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for match_id_candidates_rcpp.
}
