\name{monoiosotopic_mass}
\alias{monoiosotopic_mass}
\title{Monoiosotopic Mass}
\description{
Given a vector of string formulas, return a vector of compound monoisotopic masses. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
monoiosotopic_mass(formula, num_digits, debug)
}
\arguments{
\item{formula}{Argument formula.}

\item{num_digits}{Argument num_digits.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for monoiosotopic_mass.
}
