\name{name_summaries}
\alias{name_summaries}
\title{Name Summaries}
\description{
Given a vector of string formulas, return a vector of compound monoisotopic masses. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
name_summaries(compoundName, debug)
}
\arguments{
\item{compoundName}{Argument compoundName.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for name_summaries.
}
