\name{peptide_sequence_to_formula}
\alias{peptide_sequence_to_formula}
\title{Peptide Sequence To Formula}
\description{
Convert a peptide sequence to its molecular formula string. It takes the provided parameters and performs the corresponding mass spectrometry analysis step.
}
\usage{
peptide_sequence_to_formula(peptideSequence, verbose)
}
\arguments{
\item{peptideSequence}{Argument peptideSequence.}

\item{verbose}{Argument verbose.}

}
\value{
Returns String.
}
\details{
Generated documentation for peptide_sequence_to_formula.
}
