\name{qqq_peaks}
\alias{qqq_peaks}
\title{Qqq Peaks}
\description{
rt types quant types smoothed quant types Parameters always need to compute bounds to pass back complete information Samples rt types quant types smoothed quant types rt types quant types smoothed quant types rt types quant types smoothed quant types rt types quant types smoothed quant types
}
\usage{
qqq_peaks(mzml_file, query_transitions, params, debug, verbose)
}
\arguments{
\item{mzml_file}{A numeric value specifying a mass-to-charge (m/z) ratio.}

\item{query_transitions}{A parameter specifying the query transitions.}

\item{params}{A parameter specifying the params.}

\item{debug}{A logical value indicating whether to enable detailed debugging output.}

\item{verbose}{A logical value indicating whether to print progress and informative messages.}

}
\value{
A data frame containing the results of the mass spectrometry analysis or data extraction.
}
\details{
This function is part of the Rcpp-based backend for the mzkit package. It performs operations related to qqq peaks.
The implementation is written in C++ for performance and exposed to R via Rcpp.
}
