\name{repick_peaks}
\alias{repick_peaks}
\title{Repick Peaks}
\description{
DataFrame with instructions for repicking. MUST BE SORTED by (sampleId, groupId) - otherwise, this function will not work.
}
\usage{
repick_peaks(peaks_to_repick, encodedParams, verbose, debug)
}
\arguments{
\item{peaks_to_repick}{Argument peaks_to_repick.}

\item{encodedParams}{Argument encodedParams.}

\item{verbose}{Argument verbose.}

\item{debug}{Argument debug.}

}
\value{
Returns DataFrame.
}
\details{
Generated documentation for repick_peaks.
}
